delete de curso adicionado

+++ CoursesService/CoursesService/Repositories/CourseRepository.cs
@@ -25,10 +25,14 @@ namespace CoursesService.Repositories
             _context.Courses.Update(course);
             await _context.SaveChangesAsync();
         }
-        public async Task DeleteAsync(Course course)
+        public async Task DeleteAsync(int id)
         {
-            _context.Courses.Remove(course);
-            await _context.SaveChangesAsync();
+            var course = await _context.Courses.FindAsync(id);
+            if (course != null)
+            {
+                _context.Courses.Remove(course);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
